+++ typescript/src/index.ts
@@ -1,2 +1,6 @@
+export { KurtosisContext } from "./lib/kurtosis_context/kurtosis_context";
+export { EnclaveContext } from "./lib/enclave_context/enclave_context";
+export { ApiContainerContext } from "./lib/api_container_context/api_container_context"
+
 export { LISTEN_PORT, LISTEN_PROTOCOL } from "./kurtosis_engine_rpc_api_consts/kurtosis_engine_rpc_api_consts";
 export { KURTOSIS_ENGINE_API_VERSION } from "./kurtosis_engine_api_version/kurtosis_engine_api_version";