+++ typescript/src/lib/kurtosis_context/kurtosis_context.ts
@@ -3,11 +3,11 @@ import { LISTEN_PORT} from "../../kurtosis_engine_rpc_api_consts/kurtosis_engine
 import {EnclaveContext} from "../enclave_context/enclave_context";
 import * as grpc from "grpc";
 import { Result, err, ok } from "neverthrow";
-import {newCreateEnclaveArgs, newDestroyEnclaveArgs} from "../constructor_calls";
+import {newCreateEnclaveArgs, newDestroyEnclaveArgs, newStopEnclaveArgs} from "../constructor_calls";
 import {
     CreateEnclaveArgs,
     CreateEnclaveResponse, DestroyEnclaveArgs, EnclaveAPIContainerInfo, EnclaveAPIContainerStatusMap,
-    EnclaveInfo, GetEnclavesResponse
+    EnclaveInfo, GetEnclavesResponse, StopEnclaveArgs
 } from "../../kurtosis_engine_rpc_api_bindings/engine_service_pb";
 import {ApiContainerContext} from "../api_container_context/api_container_context";
 import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
@@ -22,6 +22,7 @@ export class KurtosisContext {
         this.client = client;
     }
 
+    // Attempts to create a KurtosisContext connected to a Kurtosis engine running locally
     public static newKurtosisContextFromLocalEngine(): Result<KurtosisContext, Error>{
         const kurtosisEngineSocketStr: string = `${LOCAL_HOST_IP_ADDRESS_STR}:${LISTEN_PORT}`;
 
@@ -115,6 +116,26 @@ export class KurtosisContext {
         const enclavesMap: Map<string, EnclaveContext> = this.newEnclaveContextMapFromEnclaveInfoMap(response.getEnclaveInfoMap())
 
         return ok(enclavesMap)
+    }
+
+    public async stopEnclave(enclaveId: string): Promise<Result<null, Error>> {
+        const args: StopEnclaveArgs = newStopEnclaveArgs(enclaveId)
+
+        const stopEnclavePromise: Promise<Result<null, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.stopEnclave(args, (error: Error | null, _unusedResponse?: google_protobuf_empty_pb.Empty) => {
+                if (error === null) {
+                    resolve(ok(null));
+                } else {
+                    resolve(err(error));
+                }
+            })
+        });
+        const stopEnclaveResult: Result<null, Error> = await stopEnclavePromise;
+        if (!stopEnclaveResult.isOk()) {
+            return err(stopEnclaveResult.error);
+        }
+
+        return ok(null);
     }
 
     public async destroyEnclave(enclaveId: string): Promise<Result<null, Error>> {