+++ docs/changelog.md
@@ -1,4 +1,15 @@
 # TBD
+### Features
+* Added `KurtosisContext.StopEnclave`
+* All Go `KurtosisContext` methods now accept a `Context`
+
+### Fixes
+* Fixed panic that would happen when getting enclave info
+* Exports `KurtosisContext`, `EnclaveContext`, and `ApiContainerContext` from the Typescript library
+
+### Breaking Changes
+* All `KurtosisContext` methods in the Go library accept a `Context`
+    * Users should pass in whichever context is appropriate
 
 # 0.6.0
 ### Features